+++ terraform/cicd-terraform-aws-codecommit.tf
@@ -1,4 +1,4 @@
-### Golang Terraform Pipeline (GitHub, AWS CodeBuild, AWS Pipeline, Amazon SNS) ###
+### Golang Terraform Pipeline (AWS CodeCommit, AWS CodeBuild, AWS Pipeline, Amazon SNS) ###
 
 # This script includes:
 #  - ECR repository creation
@@ -8,7 +8,7 @@
 #  - ECS task definition and service creation for staging environment
 #  - CloudWatch Event rule and target for manual approval notification
 #  - Outputs for the staging and production URLs, build number, and semantic versions for staging and production.
-#  - Using custom image AWS CodeBuild `devopscorner/cicd:4.0`
+#  - Using custom image AWS CodeBuild `devopscorner/cicd:codebuild-4.0`
 
 terraform {
   required_version = ">= 1.0.9"
@@ -28,7 +28,7 @@ provider "aws" {
 data "aws_caller_identity" "current" {}
 
 resource "aws_ecr_repository" "bookstore" {
-  name = "bookstore"
+  name = "devopscorner/bookstore"
 }
 
 resource "aws_sns_topic" "bookstore-topic" {
@@ -38,7 +38,7 @@ resource "aws_sns_topic" "bookstore-topic" {
 resource "aws_sns_topic_subscription" "bookstore-subscription" {
   topic_arn = aws_sns_topic.bookstore-topic.arn
   protocol  = "email"
-  endpoint  = "youremail@example.com"
+  endpoint  = "support@devopscorner.id"
 }
 
 resource "aws_codebuild_project" "bookstore-prod" {
@@ -52,7 +52,7 @@ resource "aws_codebuild_project" "bookstore-prod" {
   environment {
     compute_type = "BUILD_GENERAL1_SMALL"
     # image      = "aws/codebuild/standard:5.0"
-    image = "devopscorner/cicd:4.0"
+    image = "devopscorner/cicd:codebuild-4.0"
     type  = "LINUX_CONTAINER"
     environment_variable {
       name  = "AWS_REGION"
@@ -68,13 +68,10 @@ resource "aws_codebuild_project" "bookstore-prod" {
     }
   }
   source {
-    type            = "GITHUB"
-    location        = "https://github.com/devopscorner/golang-deployment"
+    type            = "CODECOMMIT"
+    location        = "your-repo-name"
     git_clone_depth = 1
     buildspec       = file("${path.module}/buildspec.yaml")
-    auth {
-      type = "OAUTH"
-    }
   }
 }
 
@@ -89,7 +86,7 @@ resource "aws_codebuild_project" "bookstore-staging" {
   environment {
     compute_type = "BUILD_GENERAL1_SMALL"
     # image      = "aws/codebuild/standard:5.0"
-    image = "devopscorner/cicd:4.0"
+    image = "devopscorner/cicd:codebuild-4.0"
     type  = "LINUX_CONTAINER"
     environment_variable {
       name  = "AWS_REGION"
@@ -105,13 +102,10 @@ resource "aws_codebuild_project" "bookstore-staging" {
     }
   }
   source {
-    type            = "GITHUB"
-    location        = "https://github.com/devopscorner/golang-deployment"
+    type            = "CODECOMMIT"
+    location        = "your-repo-name"
     git_clone_depth = 1
     buildspec       = file("${path.module}/buildspec.yaml")
-    auth {
-      type = "OAUTH"
-    }
   }
 }
 
@@ -121,7 +115,6 @@ resource "aws_codepipeline" "bookstore" {
   artifact_store {
     type     = "S3"
     location = "your-artifact-store-bucket"
-
     encryption_key {
       type = "KMS"
       id   = "your-kms-key-id"
@@ -134,14 +127,12 @@ resource "aws_codepipeline" "bookstore" {
       name             = "Source"
       category         = "Source"
       owner            = "AWS"
-      provider         = "GitHub"
+      provider         = "CodeCommit"
       version          = "1"
       output_artifacts = ["SourceArtifact"]
       configuration = {
-        Owner      = "your-username"
-        Repo       = "your-repo"
-        Branch     = "main"
-        OAuthToken = "your-github-token"
+        RepositoryName = "your-repo-name"
+        BranchName     = "main"
       }
     }
   }
@@ -188,7 +179,7 @@ resource "aws_codepipeline" "bookstore" {
       version         = "1"
       input_artifacts = ["BuildArtifact"]
       configuration = {
-        ClusterName = "my-ecs-cluster"
+        ClusterName = "bookstore-cluster"
         ServiceName = "bookstore-staging"
         FileName    = "imagedefinitions.json"
         ImageUri    = "${aws_ecr_repository.bookstore.repository_url}:${var.semver_staging}"
@@ -206,7 +197,7 @@ resource "aws_codepipeline" "bookstore" {
       version         = "1"
       input_artifacts = ["BuildArtifact"]
       configuration = {
-        ClusterName = "my-ecs-cluster"
+        ClusterName = "bookstore-cluster"
         ServiceName = "bookstore-prod"
         FileName    = "imagedefinitions.json"
         ImageUri    = "${aws_ecr_repository.bookstore.repository_url}:${var.semver_prod}"
@@ -222,22 +213,21 @@ resource "aws_codepipeline" "bookstore" {
 
 resource "aws_ecs_task_definition" "bookstore" {
   family = "bookstore"
-  container_definitions = jsonencode([
-    {
-      name               = "bookstore"
-      image              = "${aws_ecr_repository.bookstore.repository_url}:${var.semver_staging}"
-      cpu                = 512
-      memory_reservation = 512
-      port_mappings = {
-        container_port = 8080
-        host_port      = 0
-      }
+  container_definitions = jsonencode([{
+    name               = "bookstore"
+    image              = "${aws_ecr_repository.bookstore.repository_url}:${var.semver_staging}"
+    cpu                = 512
+    memory_reservation = 512
+    port_mappings = {
+      container_port = 8080
+      host_port      = 0
+    }
   }])
 }
 
 resource "aws_ecs_service" "bookstore-staging" {
   name            = "bookstore-staging"
-  cluster         = "my-ecs-cluster"
+  cluster         = "bookstore-cluster"
   task_definition = aws_ecs_task_definition.bookstore.arn
   desired_count   = 2
   launch_type     = "EC2"