+++ terraform/cicd-terraform-github.tf
@@ -1,36 +1,15 @@
-# Golang Deployment - CI/CD with Terraform AWS CodeBuild, AWS CodePipeline & Amazon SNS
+### Golang Terraform Pipeline (GitHub, AWS CodeBuild, AWS Pipeline, Amazon SNS) ###
 
-Kubernetes Deployment for Simple Golang API
+# This script includes:
+#  - ECR repository creation
+#  - SNS topic creation and subscription
+#  - CodeBuild project creation for production and staging environments
+#  - CodePipeline creation with four stages: Source, Build, ManualApproval, and Deploy (two times)
+#  - ECS task definition and service creation for staging environment
+#  - CloudWatch Event rule and target for manual approval notification
+#  - Outputs for the staging and production URLs, build number, and semantic versions for staging and production.
+#  - Using custom image AWS CodeBuild `devopscorner/cicd:codebuild-4.0`
 
-![goreport](https://goreportcard.com/badge/github.com/devopscorner/golang-deployment)
-![all contributors](https://img.shields.io/github/contributors/devopscorner/golang-deployment)
-![tags](https://img.shields.io/github/v/tag/devopscorner/golang-deployment?sort=semver)
-[![docker pulls](https://img.shields.io/docker/pulls/devopscorner/bookstore.svg)](https://hub.docker.com/r/devopscorner/bookstore/)
-![download all](https://img.shields.io/github/downloads/devopscorner/golang-deployment/total.svg)
-![view](https://views.whatilearened.today/views/github/devopscorner/golang-deployment.svg)
-![clone](https://img.shields.io/badge/dynamic/json?color=success&label=clone&query=count&url=https://github.com/devopscorner/golang-deployment/blob/master/clone.json?raw=True&logo=github)
-![issues](https://img.shields.io/github/issues/devopscorner/golang-deployment)
-![pull requests](https://img.shields.io/github/issues-pr/devopscorner/golang-deployment)
-![forks](https://img.shields.io/github/forks/devopscorner/golang-deployment)
-![stars](https://img.shields.io/github/stars/devopscorner/golang-deployment)
-[![license](https://img.shields.io/github/license/devopscorner/golang-deployment)](https://img.shields.io/github/license/devopscorner/golang-deployment)
-
----
-
-## Example CI/CD Terraform Script `cicd-terraform.tf`
-
-### This script includes:
-
-  - ECR repository creation
-  - SNS topic creation and subscription
-  - CodeBuild project creation for production and staging environments
-  - CodePipeline creation with four stages: Source, Build, ManualApproval, and Deploy (two times)
-  - ECS task definition and service creation for staging environment
-  - CloudWatch Event rule and target for manual approval notification
-  - Outputs for the staging and production URLs, build number, and semantic versions for staging and production.
-  - Using custom image AWS CodeBuild `devopscorner/cicd:codebuild-4.0`
-
-```
 terraform {
   required_version = ">= 1.0.9"
 
@@ -49,7 +28,7 @@ provider "aws" {
 data "aws_caller_identity" "current" {}
 
 resource "aws_ecr_repository" "bookstore" {
-  name = "bookstore"
+  name = "devopscorner/bookstore"
 }
 
 resource "aws_sns_topic" "bookstore-topic" {
@@ -59,7 +38,7 @@ resource "aws_sns_topic" "bookstore-topic" {
 resource "aws_sns_topic_subscription" "bookstore-subscription" {
   topic_arn = aws_sns_topic.bookstore-topic.arn
   protocol  = "email"
-  endpoint  = "youremail@example.com"
+  endpoint  = "support@devopscorner.id"
 }
 
 resource "aws_codebuild_project" "bookstore-prod" {
@@ -73,8 +52,8 @@ resource "aws_codebuild_project" "bookstore-prod" {
   environment {
     compute_type = "BUILD_GENERAL1_SMALL"
     # image      = "aws/codebuild/standard:5.0"
-    image        = "devopscorner/cicd:codebuild-4.0"
-    type         = "LINUX_CONTAINER"
+    image = "devopscorner/cicd:codebuild-4.0"
+    type  = "LINUX_CONTAINER"
     environment_variable {
       name  = "AWS_REGION"
       value = "ap-southeast-1"
@@ -110,8 +89,8 @@ resource "aws_codebuild_project" "bookstore-staging" {
   environment {
     compute_type = "BUILD_GENERAL1_SMALL"
     # image      = "aws/codebuild/standard:5.0"
-    image        = "devopscorner/cicd:codebuild-4.0"
-    type         = "LINUX_CONTAINER"
+    image = "devopscorner/cicd:codebuild-4.0"
+    type  = "LINUX_CONTAINER"
     environment_variable {
       name  = "AWS_REGION"
       value = "ap-southeast-1"
@@ -142,7 +121,6 @@ resource "aws_codepipeline" "bookstore" {
   artifact_store {
     type     = "S3"
     location = "your-artifact-store-bucket"
-
     encryption_key {
       type = "KMS"
       id   = "your-kms-key-id"
@@ -209,7 +187,7 @@ resource "aws_codepipeline" "bookstore" {
       version         = "1"
       input_artifacts = ["BuildArtifact"]
       configuration = {
-        ClusterName = "my-ecs-cluster"
+        ClusterName = "bookstore-cluster"
         ServiceName = "bookstore-staging"
         FileName    = "imagedefinitions.json"
         ImageUri    = "${aws_ecr_repository.bookstore.repository_url}:${var.semver_staging}"
@@ -227,7 +205,7 @@ resource "aws_codepipeline" "bookstore" {
       version         = "1"
       input_artifacts = ["BuildArtifact"]
       configuration = {
-        ClusterName = "my-ecs-cluster"
+        ClusterName = "bookstore-cluster"
         ServiceName = "bookstore-prod"
         FileName    = "imagedefinitions.json"
         ImageUri    = "${aws_ecr_repository.bookstore.repository_url}:${var.semver_prod}"
@@ -243,22 +221,21 @@ resource "aws_codepipeline" "bookstore" {
 
 resource "aws_ecs_task_definition" "bookstore" {
   family = "bookstore"
-  container_definitions = jsonencode([
-    {
-      name               = "bookstore"
-      image              = "${aws_ecr_repository.bookstore.repository_url}:${var.semver_staging}"
-      cpu                = 512
-      memory_reservation = 512
-      port_mappings = {
-        container_port = 8080
-        host_port      = 0
-      }
+  container_definitions = jsonencode([{
+    name               = "bookstore"
+    image              = "${aws_ecr_repository.bookstore.repository_url}:${var.semver_staging}"
+    cpu                = 512
+    memory_reservation = 512
+    port_mappings = {
+      container_port = 8080
+      host_port      = 0
+    }
   }])
 }
 
 resource "aws_ecs_service" "bookstore-staging" {
   name            = "bookstore-staging"
-  cluster         = "my-ecs-cluster"
+  cluster         = "bookstore-cluster"
   task_definition = aws_ecs_task_definition.bookstore.arn
   desired_count   = 2
   launch_type     = "EC2"
@@ -308,4 +285,3 @@ output "bookstore-semver-staging" {
 output "bookstore-semver-prod" {
   value = var.semver_prod
 }
-```