+++ CHANGELOG.md
@@ -1,4 +1,48 @@
-## Changelog GO App
+# Golang Deployment
+
+Kubernetes Deployment for Simple Golang API
+
+![goreport](https://goreportcard.com/badge/github.com/devopscorner/golang-deployment) ![all contributors](https://img.shields.io/github/contributors/devopscorner/golang-deployment) ![tags](https://img.shields.io/github/v/tag/devopscorner/golang-deployment?sort=semver) [![docker pulls](https://img.shields.io/docker/pulls/devopscorner/bookstore.svg)](https://hub.docker.com/r/devopscorner/bookstore/) ![download all](https://img.shields.io/github/downloads/devopscorner/golang-deployment/total.svg) ![download latest](https://img.shields.io/github/downloads/devopscorner/golang-deployment/3.1/total) ![view](https://views.whatilearened.today/views/github/devopscorner/golang-deployment.svg) ![clone](https://img.shields.io/badge/dynamic/json?color=success&label=clone&query=count&url=https://github.com/devopscorner/golang-deployment/blob/master/clone.json?raw=True&logo=github) ![issues](https://img.shields.io/github/issues/devopscorner/golang-deployment) ![pull requests](https://img.shields.io/github/issues-pr/devopscorner/golang-deployment) ![forks](https://img.shields.io/github/forks/devopscorner/golang-deployment) ![stars](https://img.shields.io/github/stars/devopscorner/golang-deployment) [![license](https://img.shields.io/github/license/devopscorner/golang-deployment)](https://img.shields.io/github/license/devopscorner/golang-deployment)
+
+## Available Tags
+
+### Alpine
+
+| Image name | Size |
+|------------|------|
+| `devopscorner/bookstore:latest` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/latest.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=latest) |
+| `devopscorner/bookstore:alpine` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/alpine.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=alpine) |
+| `devopscorner/bookstore:alpine-latest` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/alpine-latest.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=alpine-latest) |
+| `devopscorner/bookstore:alpine-3.15` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/alpine-3.15.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=alpine-3.15) |
+| `devopscorner/bookstore:go1.19-alpine3.15` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/go1.19-alpine3.15.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=go1.19-alpine3.15) |
+| `devopscorner/bookstore:go1.19.3-alpine3.15` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/go1.19.3-alpine3.15.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=go1.19.3-alpine3.15) |
+| `devopscorner/bookstore:alpine-3.16` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/alpine-3.16.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=alpine-3.16) |
+| `devopscorner/bookstore:go1.19-alpine3.16` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/go1.19-alpine3.16.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=go1.19-alpine3.16) |
+| `devopscorner/bookstore:go1.19.5-alpine3.16` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/go1.19.5-alpine3.16.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=go1.19.5-alpine3.16) |
+| `devopscorner/bookstore:alpine-3.17` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/alpine-3.17.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=alpine-3.17) |
+| `devopscorner/bookstore:go1.19-alpine3.17` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/go1.19-alpine3.17.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=go1.19-alpine3.17) |
+| `devopscorner/bookstore:go1.19.5-alpine3.17` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/go1.19.5-alpine3.17.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=go1.19.5-alpine3.17) |
+
+
+### Alpine (Depreciated)
+| Image name | Size |
+|------------|------|
+| `devopscorner/bookstore:go1.18-alpine3.15` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/go1.18-alpine3.15.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=go1.18-alpine3.15) |
+| `devopscorner/bookstore:go1.18-alpine3.16` | [![docker image size](https://img.shields.io/docker/image-size/devopscorner/bookstore/go1.18-alpine3.16.svg?label=Image%20size&logo=docker)](https://hub.docker.com/repository/docker/devopscorner/bookstore/tags?page=1&ordering=last_updated&name=go1.18-alpine3.16) |
+
+---
+
+### version 3.3
+
+- All features in version 3.2
+- Added Terraform script:
+  - GitHub, AWS CodeBuild, AWS CodePipeline & Amazon SNS
+  - AWS CodeCommit, AWS CodeBuild, AWS CodePipeline & Amazon SNS
+- Added CloudFormation script:
+  - GitHub, AWS CodeBuild, AWS CodePipeline & Amazon SNS
+  - AWS CodeCommit, AWS CodeBuild, AWS CodePipeline & Amazon SNS
+
+---
 
 ### version 3.3
 
@@ -20,6 +64,8 @@
   - **Terraform AWS CodeBuild, AWS CodePipeline & Amazon SNS**, detail [here](docs/deployment-terraform.md) link
   - **TravisCI**, detail [here](docs/deployment-travisci.md) link
 
+---
+
 ### version 3.2
 
 - All features in version 3.1
@@ -30,6 +76,8 @@
   - GitHub Action (`cicd-github.yml`)
   - GitLab (`cicd-gitlab.yml`)
   - Jenkins CI/CD (`cicd-jenkins.jenkinsfile`)
+
+---
 
 ### version 3.1
 
@@ -51,6 +99,8 @@
 - Refactoring build, tag, push & pull script for ECR
 - Refactoring `makefile` script automation for build, tag, push & pull
 
+---
+
 ### version 3.0
 
 - All features in version 2.3
@@ -61,10 +111,14 @@
 - Refactoring source code (moving) dependencies to `devopscorner/golang-deployment`
 - Update `gorm` model & sqlite connection driver
 
+---
+
 ### version 2.3
 
 - All features in version 2.2
 - Refactoring path & references docs
+
+---
 
 ### version 2.2
 
@@ -74,6 +128,8 @@
 - Add documentation for build, tag & push container image to **DockerHub**, go to [this](docs/container-bookstore-dockerhub.md) link
 - Add documentation for build, tag & push container to **Amazon ECR (Elastic Container Registry)**, go to [this](docs/container-bookstore-ecr.md) link
 - Refactoring workflow documentation, go to [this](docs/workflow-cicd-bookstore-pipeline.md) link
+
+---
 
 ### version 2.1
 
@@ -89,6 +145,8 @@
   - Azure DevSecOps Pipeline
     ![Azure DevSecOps Pipeline](docs/assets/gitops-devsecops-azure.png)
 
+---
+
 ### version 2.0
 
 - IAM Role sample for CodeBuild & CodePipeline
@@ -100,6 +158,8 @@
 - Setup `~/.ssh/config` for authorization config ssh key 3rd party repository
 - Dynamic Tags with COMMIT_HASH
 
+---
+
 ### version 1.0
 
 - Golang API Rest (bookstore)
@@ -108,6 +168,8 @@
 - Push Container to ECR
 - Deploy Kubernetes with Helm Values
 - Buildspec for AWS CodeBuild & AWS CodePipeline
+
+---
 
 ### version 0.1